Get react-native-navigation as arg instead of importing
react-native-navigation unfortunately has side effects on import, such
as registering a native component for SharedTransitionElement, and this
side effect cannot happen twice. This is why we delegate the application
project to do the import once, and pass it down to this library.

This is a breaking change to the API